+++ test/rcc-redis-adapter.test.js
@@ -5,13 +5,13 @@ const test = require('tape');
 // Get the redis adapter
 const redis = require('../rcc-redis-adapter');
 
-const host0 = '127.0.0.1';
-const port0 = 6379;
+const host0 = redis.getDefaultHost();
+const port0 = redis.getDefaultPort();
 
 const host1 = 'localhost';
 const port1 = 9999;
 
-function addEventListeners(redisClient, desc, customOnError) {
+function addEventListeners(redisClient, desc, startMs, customOnError) {
   const [host, port] = redisClient.resolveHostAndPort();
 
   const onConnect = () => {
@@ -27,7 +27,7 @@ function addEventListeners(redisClient, desc, customOnError) {
   };
 
   const onError = err => {
-    console.log(`*** Redis client ${desc} connection to host (${host}) & port (${port}) hit ERROR ${err}`);
+    console.log(`*** Redis client ${desc} connection to host (${host}) & port (${port}) failed with ERROR - ${err}`);
     if (customOnError) {
       return customOnError(err);
     }
@@ -36,19 +36,25 @@ function addEventListeners(redisClient, desc, customOnError) {
   };
 
   const onClientError = err => {
-    console.log(`*** Redis client ${desc} connection to host (${host}) & port (${port}) hit CLIENT ERROR ${err}`);
+    console.log(`*** Redis client ${desc} connection to host (${host}) & port (${port}) failed with CLIENT ERROR - ${err}`);
     console.error(err);
   };
 
   const onEnd = () => {
+    console.log(`*** Redis client ${desc} connection to host (${host}) & port (${port}) has ENDED`);
+  };
+
+  // Not supported by `redis` module
+  const onClose = () => {
     console.log(`*** Redis client ${desc} connection to host (${host}) & port (${port}) has CLOSED`);
   };
 
-  redisClient.addEventListeners(onConnect, onReady, onReconnecting, onError, onClientError, onEnd);
+  redisClient.addEventListeners(onConnect, onReady, onReconnecting, onError, onClientError, onEnd, onClose);
 }
 
-test('createClient', t => {
+test('redis - createClient', t => {
   // Create a redis client using the redis adapter
+  let startMs = Date.now();
   const redisClient0 = redis.createClient();
   t.ok(redisClient0, `redisClient0 must exist`);
   t.notOk(redisClient0.isClosing(), `redisClient0 should not be closing yet`);
@@ -57,12 +63,13 @@ test('createClient', t => {
   t.equal(h, host0, `redisClient0 host must be ${host0}`);
   t.equal(p, port0, `redisClient0 port must be ${port0}`);
 
-  addEventListeners(redisClient0, 0, err => {
+  addEventListeners(redisClient0, 0, startMs, err => {
     t.pass(`Expected and got an error from redisClient0 (${err})`);
     redisClient0.end(true, () => {});
     t.ok(redisClient0.isClosing(), `redisClient0 should be closing now`);
   });
 
+  startMs = Date.now();
   const redisClient1 = redis.createClient({port: port0, host: host0});
   t.ok(redisClient1, `redisClient1 must exist`);
   t.notOk(redisClient1.isClosing(), `redisClient1 should not be closing yet`);
@@ -71,12 +78,13 @@ test('createClient', t => {
   t.equal(h, host0, `redisClient1 host must be ${host0}`);
   t.equal(p, port0, `redisClient1 port must be ${port0}`);
 
-  addEventListeners(redisClient1, 1, err => {
+  addEventListeners(redisClient1, 1, startMs, err => {
     t.pass(`Expected and got an error from redisClient1 (${err})`);
     redisClient1.end(true, () => {});
     t.ok(redisClient1.isClosing(), `redisClient1 should be closing now`);
   });
 
+  startMs = Date.now();
   const redisClientOptions2 = {host: host1, port: port1, string_number: true};
   const redisClient2 = redis.createClient(redisClientOptions2);
 
@@ -88,10 +96,10 @@ test('createClient', t => {
   t.equal(p, port1, `redisClient2 port must be ${port1}`);
 
   // Set and get a value for a key using the underlying `redis` module's `RedisClient` instance's methods
-  const key = 'KEY';
-  const expectedValue = 'VALUE';
+  const key = 'TEST_KEY';
+  const expectedValue = 'TEST_VALUE';
 
-  addEventListeners(redisClient2, 2, err => {
+  addEventListeners(redisClient2, 2, startMs, err => {
     t.pass(`Expected and got an error from redisClient2 (${err})`);
     redisClient2.end(true, () => {});
     t.ok(redisClient2.isClosing(), `redisClient2 should be closing now`);
@@ -99,7 +107,7 @@ test('createClient', t => {
 
   redisClient2.set(key, expectedValue, (err,) => {
     if (err) {
-      t.pass(`Expected and got a set error (${err})`);
+      t.pass(`Expected and got a set error (${err}) - ${JSON.stringify(err)} - isMovedError? ${redisClient0.getAdapter().isMovedError(err)}`);
       redisClient0.end(true);
       redisClient1.end(true);
       redisClient2.end(true);
@@ -113,7 +121,7 @@ test('createClient', t => {
         if (err) {
           t.pass(`Expected and got a get error (${err})`);
         } else {
-          t.equal(value, 'VALUE', `value must be '${expectedValue}'`);
+          t.equal(value, expectedValue, `value must be '${expectedValue}'`);
         }
         redisClient0.end(true);
         redisClient1.end(true);